requirements.txt
blinker==1.9.0
click==8.3.1
flask==3.1.2
gunicorn==23.0.0
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.3
mysql-connector-python==9.5.0
packaging==25.0
python-dotenv==1.2.1
werkzeug==3.1.5
